+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [20.0.0] - 2020-05-04
+### Added
+- Added per-group evaluation result output
+### Changed
+- Changed data structure of EvaluationResults
+
 ## [19.0.1] - 2020-04-29
 ### Added
 - Added publish to bintray functionality